separation of concern
