+++ src/services/storage.mjs
@@ -1,6 +1,4 @@
 import CryptoJS from "crypto-js";
-import { CallData, ec, encode, hash, stark } from "starknet"
-import { argentAccountClassHash } from "../utils/constants.mjs";
 
 export const StarkTradeStorage = () => {
     return {
@@ -33,21 +31,4 @@ export const sessionChecker = (ctx) => {
         return false
     }
     return true
-}
-
-export const generateAccount = (privateKey) => {
-    const starkPublicKey = ec.starkCurve.getStarkKey(privateKey);
-
-    const AConstructCalldata = CallData.compile({
-        owner: starkPublicKey,
-        guardian: '0',
-    })
-
-    const AContractAddress = hash.calculateContractAddressFromHash(
-        starkPublicKey,
-        argentAccountClassHash,
-        AConstructCalldata,
-        0
-    );
-    return AContractAddress
 }
